backend/requirements.txt
flask==3.0.0
pymongo==4.6.1
flask-jwt-extended==4.6.0
bcrypt==4.1.2
flask-cors==4.0.0
python-dotenv==1.0.0
scikit-learn==1.4.0
numpy==1.26.3
cloudinary==1.38.0
